Reorganize presenter a bit
