--- pakyow-presenter/lib/pakyow-presenter.rb
@@ -1,1 +1,0 @@
-require 'pakyow/presenter'
