+++ pakyow-presenter/lib/pakyow/presenter.rb
@@ -1,11 +1,32 @@
-require 'pakyow/support/silenceable'
-
+require "pakyow/support/silenceable"
 Pakyow::Support::Silenceable.silence_warnings do
-  require 'oga'
+  require "oga"
 end
 
-require 'pakyow/presenter/base'
-require 'pakyow/presenter/presenter'
-require 'pakyow/presenter/config/presenter'
-require 'pakyow/presenter/helpers'
-require 'pakyow/presenter/ext/app'
+require "yaml"
+
+require "pakyow/presenter/view_store"
+require "pakyow/presenter/view"
+require "pakyow/presenter/form"
+require "pakyow/presenter/template"
+require "pakyow/presenter/page"
+require "pakyow/presenter/container"
+require "pakyow/presenter/partial"
+require "pakyow/presenter/view_collection"
+require "pakyow/presenter/binder"
+require "pakyow/presenter/binder_set"
+require "pakyow/presenter/attributes"
+require "pakyow/presenter/exceptions"
+require "pakyow/presenter/view_composer"
+require "pakyow/presenter/string_doc"
+require "pakyow/presenter/string_doc_parser"
+require "pakyow/presenter/string_doc_renderer"
+require "pakyow/presenter/binding_eval"
+require "pakyow/presenter/doc_helpers"
+require "pakyow/presenter/view_version"
+require "pakyow/presenter/view_context"
+require "pakyow/presenter/view_store_loader"
+
+require "pakyow/presenter/presenter"
+require "pakyow/presenter/config"
+require "pakyow/presenter/helpers"
