+++ pakyow-presenter/lib/pakyow/presenter/config.rb
